requirements.txt
growattServer==1.7.0
requests==2.32.3
python-dotenv==1.0.1
